Move TerminalWin to interfaces sub-module.
